+++ Pod/Classes/MWZMapDelegate.h
@@ -13,6 +13,7 @@
 
 @optional
 
+- (void) mapDidLoad: (MWZMapView*) map;
 - (void) map:(MWZMapView*) map didClick:(MWZLatLon*) latlon;
 - (void) map:(MWZMapView*) map didClickOnPlace:(MWZPlace*) place;
 - (void) map:(MWZMapView*) map didClickOnVenue:(MWZVenue*) venue;
@@ -27,14 +28,6 @@
 - (void) map:(MWZMapView*) map didStartDirections: (NSString*) infos;
 - (void) map:(MWZMapView*) map didStopDirections: (NSString*) infos;
 - (void) map:(MWZMapView*) map didFailWithError: (NSError *)error;
-
-/*!
- * @brief Invoked when a main frame load completes.
- * @param map The MapView invoking the delegate method.
- * @param webView The webView has finished loading.
- */
-- (void) map:(MWZMapView *) map webViewDidFinishLoad:(WKWebView *)webView;
-
 
 @end