+++ Pod/Classes/MWZMapOptions.h
@@ -8,9 +8,12 @@
 @property MWZLatLonBounds* maxBounds;
 @property MWZLatLon* center;
 @property NSNumber* zoom;
+@property NSNumber* minZoom;
 @property NSNumber* floor;
 @property BOOL locationEnabled;
 @property BOOL beaconsEnabled;
+@property NSString* accessKey;
+@property NSString* language;
 
 - (instancetype) init;