+++ Pod/Classes/MWZMapOptions.m
@@ -5,12 +5,12 @@
 - (instancetype) init {
     self =  [super init];
     self.locationEnabled = YES;
-    self.locationEnabled = YES;
+    self.beaconsEnabled = NO;
     return self;
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat: @"MWZMapOptions: ApiKey=%@ MaxBounds=%@ Center=%@ Zoom=%@ Floor=%@ LocationEnabled=%@ BeaconsEnabled=%@", _apiKey, _maxBounds, _center, _zoom, _floor, _locationEnabled ? @"YES" : @"NO", _beaconsEnabled ? @"YES" : @"NO"];
+    return [NSString stringWithFormat: @"MWZMapOptions: ApiKey=%@ MaxBounds=%@ MinZoom=%@Center=%@ Zoom=%@ Floor=%@ LocationEnabled=%@ BeaconsEnabled=%@", _apiKey, _maxBounds, _minZoom, _center, _zoom, _floor, _locationEnabled ? @"YES" : @"NO", _beaconsEnabled ? @"YES" : @"NO"];
 }
 
 @end